"Refactor Card1 and Section1, and add 'deepmerge' dependency
The Card1 and Section1 components were refactored to improve maintainability. These components are now easier to modify with the use of typescript interfaces for their props, providing reusability.

Specifically, the Card1 component now extends Omit<CardProps, 'children'>  and has additional props such as CardHeaderProps, SubtitleProps,  and ButtonProps for extra customization. All hardcoded inline styles were moved to 'sx' props with improved CSS structure and color inheritance, which allows to easier manage styles with potential themes in the future.

In the Section1 component, the hardcoded data was moved outside of the function to avoid recreation on every re-render, improving performance. Also,