Remove unnecessary Drawer component conditional check, add overflowX style
The conditional check for displaying the Drawer component in the header was removed as it was redundant. The Drawer component itself includes the same conditional check internally. Hence, to reduce unnecessary checks and possible confusion, it has been removed.

Additionally, in the page.tsx file, the 'overflowX' property was added to the main style to prevent horizontal scrolling. This improves the user experience on smaller screens where the content may not fit perfectly.